[INFRA] Rename `default_configuration.hpp` to `default.hpp`.
