+++ include/seqan3/search/configuration/default.hpp
